Improve the fallback weights for unknown words
When generating a supervised weightlist, the number of times certain tags occur
in the corpus can be used to provide an estimate for unknown words with a matchine tag.
Example: if <n><pl> occurred m times then
we can assume that analyses of form ?*<n><pl> have a fallback weight that is somehow proportional to m.
Additionally, a default weight can be estimated as further fallback weight using laplace smoothing.